requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
numpy==1.26.0
openai==1.3.0
